+++ src/io/common.jl
@@ -1,15 +1,19 @@
 
 
 "turns top level arrays into dicts"
-function arrays_to_dicts(data::Dict{String,Any})
+function arrays_to_dicts!(data::Dict{String,Any})
     # update lookup structure
     for (k,v) in data
-        if isa(v, Array)
+        if isa(v, Array) && isa(v[1], Dict)
             #println("updating $(k)")
             dict = Dict{String,Any}()
-            for item in v
-                assert("index" in keys(item))
-                key = string(item["index"])
+            for (i,item) in enumerate(v)
+                if haskey(item, "index")
+                    key = string(item["index"])
+                else
+                    key = string(i)
+                end
+
                 if !(haskey(dict, key))
                     dict[key] = item
                 else